Parsing of board placement JSON updated to handle two players.
Known edge case: Positions are stored by using an intermediate class with two bools. If the incoming JSON only has one bool for one of the players, it still seems to be accepted when creating the resulting dictionary (likely due to the fact that we're not using a JSON schema here). I'm strongly assuming that the missing information would then default to 'false'.